chore(web): break the last recipe-form import cycle (868kvw0z3)
Fallow C. RecipeFormProps was declared in recipeForm.tsx and imported by
useRecipeForm.tsx, so the hook depended on the component that consumes it —
the last back edge in the recipeForm <-> hooks <-> services cycle set and a
layering inversion against the documented order (UI -> hooks/stores ->
services).

The type moves to a sibling types.ts that imports only from
@costwise/shared/recipe, so both the component and the hook depend on it and
neither depends on the other. Recipe, Ingredient, IngredientToDisplay and
RecipeIngredients went unused in recipeForm.tsx once the interface left, so
the import narrows to RecipeCategory.

The other six cycles this task named, and the services/ -> components/ import,
were already removed by Fallow D (868kvw157, PR #174). Fallow now reports
0 circular dependencies repo-wide, down from 7.

Acceptance criteria:
- 0 circular deps under components/recipes/** and hooks/** — 0 repo-wide
- no module under services/ imports from components/ — verified by grep
- shared form types in one place — RecipeFormProps in types.ts,
  FormFields in @costwise/shared/recipe
- pnpm build, pnpm test (129 tests), pnpm lint green

Create/edit: this is a type-only move, erased at build time. The production
build compiles the recipe create and edit routes, and the two useRecipeForm
tests cover create-mode and edit-mode submission. A signed-in browser
click-through was not run.
